// app/shared/model/product.model.ts
export class Product {
    _id: string;
    productId: string;
    productName: string;
    productDescription: string;
    price: string;
    color: string;
    productImageName: [string];
    superCategory: string;
    mainCategory: string;
    subCategory: string;
    overview: string;
    // details
    material: string;
    styleCode: string;
    skuCode: string;
    brand: string;
    dimensions: string;
    weight: string;
    assembly: string;
    packageDetails: string;
    skuCodeVerify: boolean;
}